+++ kifi-backend/modules/common/app/com/keepit/scraper/ScraperClientModule.scala
@@ -12,7 +12,6 @@ trait ScraperServiceClientModule extends ScalaModule
 
 case class ProdScraperServiceClientModule() extends ScraperServiceClientModule {
   def configure() {
-//    bind[HttpClient].to[HttpClientImpl]
   }
 
   @Singleton
