--- kifi-backend/modules/shoebox/app/com/keepit/scraper/ScraperResult.scala
@@ -1,11 +1,0 @@
-package com.keepit.scraper
-
-import com.keepit.search.Article
-import com.keepit.model.NormalizedURI
-
-sealed abstract class ScraperResult
-
-case class Scraped(article: Article, signature: Signature, redirects: Seq[HttpRedirect]) extends ScraperResult
-case class NotScrapable(destinationUrl: Option[String], redirects: Seq[HttpRedirect]) extends ScraperResult
-case class Error(httpStatusCode: Int, msg: String) extends ScraperResult
-case object NotModified extends ScraperResult
